# Swagger/api_mod.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "API",
    "description": "API for Portal System in AEM",
    "termsOfService": "None"
  },
  "host": "otvcrm01:84",
  "basePath": "/",
  "schemes": [ "http", "https" ],
  "consumes": [ "application/json" ],
  "produces": [ "application/json" ],
  "paths": {
...
    "/api/ExpectedResults": {
      "get": {
        "tags": [ "ExpectedResults" ],
        "operationId": "ApiExpectedResultsGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/ExpectedResult" }
            }
          }
        }
      },
      "post": {
        "tags": [ "ExpectedResults" ],
        "operationId": "ApiExpectedResultsPost",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "expectedResult",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/ExpectedResult" }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/ExpectedResult" }
          }
        }
      }
    },
    "/api/ExpectedResults/{id}": {
      "get": {
        "tags": [ "ExpectedResults" ],
        "operationId": "ApiExpectedResultsByIdGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/ExpectedResult" }
          }
        }
      },
      "put": {
        "tags": [ "ExpectedResults" ],
        "operationId": "ApiExpectedResultsByIdPut",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "expectedResult",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/ExpectedResult" }
          }
        ],
        "responses": { "200": { "description": "Success" } }
      },
      "delete": {
        "tags": [ "ExpectedResults" ],
        "operationId": "ApiExpectedResultsByIdDelete",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/ExpectedResult" }
          }
        }
      }
    },
    "/api/FrequentlyAskedQuestions": {
      "get": {
        "tags": [ "FrequentlyAskedQuestions" ],
        "operationId": "ApiFrequentlyAskedQuestionsGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/FrequentlyAskedQuestion" }
            }
          }
        }
      },
      "post": {
        "tags": [ "FrequentlyAskedQuestions" ],
        "operationId": "ApiFrequentlyAskedQuestionsPost",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "frequentlyAskedQuestion",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/FrequentlyAskedQuestion" }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/FrequentlyAskedQuestion" }
          }
        }
      }
    },
    "/api/FrequentlyAskedQuestions/{id}": {
      "get": {
        "tags": [ "FrequentlyAskedQuestions" ],
        "operationId": "ApiFrequentlyAskedQuestionsByIdGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/FrequentlyAskedQuestion" }
          }
        }
      },
      "put": {
        "tags": [ "FrequentlyAskedQuestions" ],
        "operationId": "ApiFrequentlyAskedQuestionsByIdPut",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "frequentlyAskedQuestion",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/FrequentlyAskedQuestion" }
          }
        ],
        "responses": { "200": { "description": "Success" } }
      },
      "delete": {
        "tags": [ "FrequentlyAskedQuestions" ],
        "operationId": "ApiFrequentlyAskedQuestionsByIdDelete",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/FrequentlyAskedQuestion" }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ExpectedResult": {
      "type": "object",
      "properties": {
        "expectedResultId": {
          "format": "uuid",
          "type": "string"
        },
        "gcimsAttributeID": { "type": "string" },
        "lang": { "type": "string" },
        "description": { "type": "string" },
        "descriptionE": { "type": "string" },
        "descriptionF": { "type": "string" },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "createdByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "FrequentlyAskedQuestion": {
      "type": "object",
      "properties": {
        "frequentlyAskedQuestionId": {
          "format": "uuid",
          "type": "string"
        },
        "descriptionE": { "type": "string" },
        "descriptionF": { "type": "string" },
        "answerE": { "type": "string" },
        "answerF": { "type": "string" },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "createdByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {}
}